Use separate output directory
