+++ .gitignore
@@ -1,7 +1,3 @@
+/dist/
 /node_modules/
 *.log
-*.js
-
-*.d.ts
-!/typings/**/*.d.ts
-
