+++ .npmignore
@@ -1,4 +1,3 @@
-!*.js
-!*.d.ts
-!/typings/*.d.ts
-
+# .gitignore without /dist/
+/node_modules/
+*.log
